Adds a 'fork me on GitHub' ribbon.
In order to gain as many quality contributions as possible, it's really
important to have one of these on your website.

There wasn't a ribbon available in the accent colour of the site so I
had to make my own.  I'm afraid my spelling isn't that great, but I had
a few of the folks in the office proof read it and they said it was
lexicographically sweet as.